+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-stellar-agent
 
+## 0.1.10
+
 ## 0.1.9
 
 ### Patch Changes