+++ packages/umbrella/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @stellar-agent-kit/all
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @stellar-agent-kit/core@0.1.10
+  - @stellar-agent-kit/adapter-mcp@0.1.10
+  - @stellar-agent-kit/plugin-anchor@0.1.10
+  - @stellar-agent-kit/plugin-asset@0.1.10
+  - @stellar-agent-kit/plugin-bridge@0.1.10
+  - @stellar-agent-kit/plugin-data@0.1.10
+  - @stellar-agent-kit/plugin-defi@0.1.10
+  - @stellar-agent-kit/plugin-defindex@0.1.10
+  - @stellar-agent-kit/plugin-domain@0.1.10
+  - @stellar-agent-kit/plugin-nft@0.1.10
+  - @stellar-agent-kit/plugin-payments@0.1.10
+  - @stellar-agent-kit/plugin-smart-wallet@0.1.10
+  - @stellar-agent-kit/plugin-soroban@0.1.10
+  - @stellar-agent-kit/plugin-trustless-work@0.1.10
+  - @stellar-agent-kit/runner@0.1.10
+
 ## 0.1.9
 
 ### Patch Changes